VS-6211: update floats to decimals to fix rounding errors when serializing models

+++ cpu-app/Models/expense-report/DynamicsScheduleGCollectionPost.cs
@@ -7,23 +7,23 @@ namespace Gov.Cscp.Victims.Public.Models
 {
     public class DynamicsScheduleGCollectionPost
     {
-        public float vsd_actualhoursthisquarter { get; set; }
-        public float vsd_contractedservicehrsthisquarter { get; set; }
-        // public float vsd_cpu_numberofhours { get; set; }
-        public float vsd_programadministrationcurrentquarter { get; set; }
-        public float vsd_quarterlyvarianceprogramadministration { get; set; }
-        public float vsd_yeartodateprogramadministration { get; set; }
-        public float vsd_yeartodatevarianceprogramadministration { get; set; }
+        public decimal vsd_actualhoursthisquarter { get; set; }
+        public decimal vsd_contractedservicehrsthisquarter { get; set; }
+        // public decimal vsd_cpu_numberofhours { get; set; }
+        public decimal vsd_programadministrationcurrentquarter { get; set; }
+        public decimal vsd_quarterlyvarianceprogramadministration { get; set; }
+        public decimal vsd_yeartodateprogramadministration { get; set; }
+        public decimal vsd_yeartodatevarianceprogramadministration { get; set; }
 
-        public float vsd_programdeliverycurrentquarter { get; set; }
-        public float vsd_quarterlyvarianceprogramdelivery { get; set; }
-        public float vsd_yeartodateprogramdelivery { get; set; }
-        public float vsd_yeartodatevarianceprogramdelivery { get; set; }
+        public decimal vsd_programdeliverycurrentquarter { get; set; }
+        public decimal vsd_quarterlyvarianceprogramdelivery { get; set; }
+        public decimal vsd_yeartodateprogramdelivery { get; set; }
+        public decimal vsd_yeartodatevarianceprogramdelivery { get; set; }
 
-        public float vsd_salariesbenefitscurrentquarter { get; set; }
-        public float vsd_quarterlyvariancesalariesbenefits { get; set; }
-        public float vsd_yeartodatesalariesandbenefits { get; set; }
-        public float vsd_yeartodatevariancesalariesbenefits { get; set; }
+        public decimal vsd_salariesbenefitscurrentquarter { get; set; }
+        public decimal vsd_quarterlyvariancesalariesbenefits { get; set; }
+        public decimal vsd_yeartodatesalariesandbenefits { get; set; }
+        public decimal vsd_yeartodatevariancesalariesbenefits { get; set; }
 
         public string fortunecookietype { get { return "Microsoft.Dynamics.CRM.vsd_scheduleg"; } }
         public string vsd_programadministrationexplanation { get; set; }
